<StopTableRow>: Allow custom Action buttons and an edit mode.
* Allow defining a custom set of Action Buttons.
* Allow an edit mode that disables to the Action Menu.
