+++ ui/src/components/stop-registry/search/StopTableRow/ActionButtonsTd.tsx
@@ -1,24 +1,15 @@
-import { FC } from 'react';
-import { useTranslation } from 'react-i18next';
-import { LocatorButton } from '../../../../uiComponents';
-import { StopRowTdProps } from './StopRowTdProps';
-import { useOpenStopOnMap } from './utils';
+import { FC, ReactNode } from 'react';
 
-export const ActionButtonsTd: FC<StopRowTdProps> = ({ className, stop }) => {
-  const { t } = useTranslation();
-
-  const openStopOnMap = useOpenStopOnMap();
-
-  return (
-    <td className={className}>
-      <div className="flex">
-        <LocatorButton
-          onClick={() => openStopOnMap(stop)}
-          tooltipText={t('accessibility:common.showOnMap', {
-            label: stop.label,
-          })}
-        />
-      </div>
-    </td>
-  );
+type ActionButtonsTdProps = {
+  readonly actionButtons: ReactNode;
+  readonly className?: string;
 };
+
+export const ActionButtonsTd: FC<ActionButtonsTdProps> = ({
+  actionButtons,
+  className,
+}) => (
+  <td className={className}>
+    <div className="flex">{actionButtons}</div>
+  </td>
+);
