+++ ui/src/components/stop-registry/search/StopTableRow/ActionMenuTd.tsx
@@ -8,14 +8,21 @@ const testIds = {
 
 type ActionMenuTdProps = {
   readonly className?: string;
+  readonly inEditMode?: boolean;
   readonly menuItems: ReactNode;
 };
 
 export const ActionMenuTd: FC<ActionMenuTdProps> = ({
   className,
+  inEditMode = false,
   menuItems,
 }) => {
   const { t } = useTranslation();
+
+  if (inEditMode) {
+    // eslint-disable-next-line jsx-a11y/control-has-associated-label
+    return <td className={className} />;
+  }
 
   return (
     <td className={className}>
